+++ tests/testthat/test-modelplot.R
@@ -1,4 +1,4 @@
-skip_on_cran()
+# skip_on_cran()
 
 test_that("single model", {